Updated for first draft of chapters 14 to 21

+++ PracticalApps/NorthwindML/Models/CartItem.cs
@@ -2,8 +2,7 @@ namespace NorthwindML.Models
 {
   public class CartItem
   {
-    public int ProductID { get; set; }
-    
+    public long ProductID { get; set; }
     public string ProductName { get; set; }
   }
 }
